requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
google-cloud==0.34.0
google-cloud-secret-manager==2.4.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.37.1
httplib2==0.19.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
libcst==0.3.18
MarkupSafe==1.1.1
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.14.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
gunicorn==20.0.4
